Make process-props function
This function process the props that will be passed to the component.
This initial implementation of process-props adds support for taking
any prop whose type is `component`, doing a lookup in the component
library for that component, then adding that component as a prop to be
passed into the rendered component.

For example, the `NextLink` component will be passed into the
AvailableWorkflowsFromEndpoint component as the prop `link`.

```yaml
view:
  component: AvailableWorkflowsFromEndpoint
  props:
    link:
      type: component
      value: NextLink
```