Add AvailableWorkflowsFromEndpoint component to demo-app OT components
This is almost an identical copy of the AvailableWorkflowsFromEndpoint
component in the demo-app. The only difference is it removes the
"next/link" dependency. Instead we declare a Component Shim called
CSLinkShape and use that in AvailableWorkflowsFromEndpoint.

CSLinkShape declares the interface for the component / function
AvailableWorkflowsFromEndpoint depends on and expects it to be
provided the application.