Update rendering code to create signal store in globalContext
This change checks all component props for a given workflow.
If they are of type "Signal", we create a new preact signal and
save into the globalContext object. When we render the component,
we pass it any signals it needs. This allows components to read
and update state globally, allowing easy data sharing between
components.

This change also creates the concept of signal types for components.
In the future this should make it easier to dynamically show which
components can be linked with other components by virtue of sharing
the same props interface. e.g. a form component that modifies the
AccountID signal can be linked with a display component that reads
an AccountID signal.