Refactor OTUqiDataProvider to use signals hooks
There was some weird interaction with react hooks
and signals leading to failures. I think signals
uses react hooks underneath the hood and using
the hooks again inside of components causes weirdness.

As a general rule, when we want to react to signals
updates we should probably use the signals hooks
useComputed useSignalEffect instead of the react
hooks.