Expose useSignals hook
Since we are not using the babel compiler to
transform static calls to signals code to be
reactive, the docs recommend that useSignals
be invoked for every component you want to be
reactive. i.e. any component that you want to
be rerendered if the referenced signal is updated.
This means that all our components will need to
call useSignals().

Alternatively, we can add babel to our build
step, but some clients may have less flexibility
to have this as a dependency.

https://www.npmjs.com/package/@preact/signals-react